feat: implement Cycles context provider and countdown component with styled components
