+++ src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,0 +1,45 @@
+import { FormContainer, MinutesAmountInput, TaskInput } from "./styles";
+
+import { useContext } from "react";
+import { CyclesContext } from "../../../../providers/CyclesProvider";
+import { useFormContext } from "react-hook-form";
+
+export function NewCycleForm() {
+  const { activeCycle } = useContext(CyclesContext);
+  const { register } = useFormContext();
+
+  return (
+    <FormContainer>
+      <label htmlFor="task">Vou trabalhar em:</label>
+      <TaskInput
+        id="task"
+        type="text"
+        list="tasks-suggestions"
+        placeholder="Dê um nome para o seu projeto"
+        disabled={!!activeCycle}
+        {...register("task")}
+      />
+
+      <datalist id="tasks-suggestions">
+        <option value="Projeto 1" />
+        <option value="Projeto 2" />
+        <option value="Projeto 3" />
+        <option value="Banana" />
+      </datalist>
+
+      <label htmlFor="minutesAmount">durante</label>
+      <MinutesAmountInput
+        id="minutesAmount"
+        type="number"
+        placeholder="00"
+        step={5}
+        min={5}
+        max={60}
+        disabled={!!activeCycle}
+        {...register("minutesAmount", { valueAsNumber: true })}
+      />
+
+      <span>minutos.</span>
+    </FormContainer>
+  );
+}
