+++ src/providers/CyclesProvider.tsx
@@ -1,0 +1,12 @@
+import { createContext } from "react";
+import { Cycle } from "../@types/types";
+
+interface CyclesContextData {
+  activeCycle: Cycle | undefined;
+  activeCycleId: string | null;
+  markCurrentCycleAsFinished: () => void;
+  amountSecondsPassed: number;
+  setSecondsPassed: (seconds: number) => void;
+}
+
+export const CyclesContext = createContext({} as CyclesContextData);
